Prepared project for multi page developing
