+++ src/Pages/App.js
@@ -1,0 +1,12 @@
+import './App.css';
+import React from 'react';
+import {Login} from './Login/Login.js';
+
+
+function App() {
+    return(
+      <Login/>
+    );
+}
+
+export default App;
